Move ServicesNav to own component
