+++ src/Components/index.js
@@ -3,4 +3,5 @@ export { default as CustomFooter } from "./CustomFooter";
 export { default as PageHeader } from "./PageHeader";
 export { default as FooterList } from "./FooterList";
 export { default as Service } from "./Service";
+export { default as CustomServicesNav } from "./CustomServicesNav";
 export { default as Project } from "./Project";
